🚚 Move `get_cached_json` to cache_hooks.py
